fix: export only the plugin factory from the plugin entry point
opencode calls every exported function in a plugin module as a plugin
factory and installs each return value as hooks. src/index.ts exported
four helpers alongside the plugin, so mergePluginAgents was being invoked
as (input, undefined) and threw on Object.entries, which opencode
swallowed while still exiting 0:

  failed to load plugin ... Object.entries requires that input parameter
  not be null or undefined

Hook registration survived only because ESM namespace keys sort
alphabetically and OpenCodeConfigPlugin starts with a capital letter,
placing it ahead of the throw. Renaming that export, or adding one
sorting earlier, would have silently dropped every hook.

Move the helpers to plugin-config.ts and leave the entry point exporting
the plugin alone.
