+++ tests/index.test.ts
@@ -2,7 +2,7 @@
 import { describe, expect, it } from "bun:test";
 import { readFile } from "node:fs/promises";
 
-import { mergePluginAgents } from "../src/index";
+import { mergePluginAgents } from "../src/plugin-config";
 
 describe("mergePluginAgents", () => {
   it("preserves user iteration limits from existing OpenCode agent config", () => {
@@ -79,5 +79,22 @@ describe("index.ts commands", () => {
     const mindmodelMatch = source.match(/mindmodel:\s*\{[^}]*agent:\s*["']([^"']+)["']/);
     expect(mindmodelMatch).not.toBeNull();
     expect(mindmodelMatch?.[1]).toBe("mm-orchestrator");
+  });
+});
+
+// opencode calls every exported function in a plugin module as a plugin factory
+// and treats each return value as a hooks object. A stray helper export is
+// therefore never harmless: it either throws and kills registration, or returns
+// something nonsensical that opencode installs as hooks. Which one happens to
+// win depends on alphabetical sort order of the module namespace, so the only
+// safe surface is the plugin itself.
+describe("plugin module surface", () => {
+  it("exports the plugin factory and nothing else callable", async () => {
+    const surface: Record<string, unknown> = await import("../src/index");
+    const callable = Object.entries(surface)
+      .filter(([, value]) => typeof value === "function")
+      .map(([name]) => name);
+
+    expect(callable).toEqual(["OpenCodeConfigPlugin"]);
   });
 });
